refactor(oc-docs): drop per-component index.ts barrels per review (BRU-3572)
Team standard (reversed from the earlier 4-file pattern): component folders
have no index.ts barrel; consumers import the concrete file directly.

- Delete the 6 barrels this PR added: Topbar/, Topbar/Brand/,
  Topbar/MobileOverflow/, IconButton/, InitialsAvatar/, OpenInBrunoButton/.
- Repoint every importer at the concrete module (e.g. '../Topbar' ->
  '../Topbar/Topbar', '../IconButton' -> '../IconButton/IconButton'),
  switching the barrel's named re-exports back to the components' default
  exports. Touches Topbar.tsx, Brand.tsx, MobileOverflow.tsx,
  OpenCollection.tsx (mount) and dev.tsx (harness).

Behavior/visual unchanged. icons split + e2e locators restructure deferred
to mirror PR #45 once it lands (icons.tsx already notes the reconciliation).
