--- packages/oc-docs/src/components/IconButton/index.ts
@@ -1,2 +1,0 @@
-export { default as IconButton } from './IconButton';
-export type { IconButtonProps } from './IconButton';
