--- packages/oc-docs/src/components/OpenInBrunoButton/index.ts
@@ -1,2 +1,0 @@
-export { default as OpenInBrunoButton } from './OpenInBrunoButton';
-export type { OpenInBrunoButtonProps } from './OpenInBrunoButton';
