--- packages/oc-docs/src/components/Topbar/MobileOverflow/index.ts
@@ -1,3 +1,0 @@
-export { default } from './MobileOverflow';
-export { default as MobileOverflow } from './MobileOverflow';
-export type { MobileOverflowProps } from './MobileOverflow';
