+++ packages/oc-docs/src/components/Topbar/Topbar.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { StyledWrapper } from './StyledWrapper';
-import Brand from './Brand';
-import MobileOverflow from './MobileOverflow';
-import { OpenInBrunoButton } from '../OpenInBrunoButton';
-import { IconButton } from '../IconButton';
+import Brand from './Brand/Brand';
+import MobileOverflow from './MobileOverflow/MobileOverflow';
+import OpenInBrunoButton from '../OpenInBrunoButton/OpenInBrunoButton';
+import IconButton from '../IconButton/IconButton';
 import { SearchIcon, HamburgerIcon } from '../../assets/icons';
 import { useTopbarLayout, showsHamburger } from '../../hooks/useTopbarLayout';
 import { useCanRunBrunoApp } from '../../hooks/useCanRunBrunoApp';
