--- packages/oc-docs/src/components/Topbar/index.ts
@@ -1,2 +1,0 @@
-export { default as Topbar } from './Topbar';
-export type { TopbarProps } from './Topbar';
